requirements.txt
six == 1.11.0
pyocr == 0.5.3
Pillow == 5.3.0